==> 00_requirements.R <==
if (require(tidyverse)) {       
  install.packages("tidyverse") #for lubridate for dealing with dates
  library(tidyverse)            
}
if (require(data.table)) {
  install.packages("data.table")
  library(data.table)
}
