==> 11_DataCleaning.Rmd <==
---
title: "STATS_107_HalfwayProj 2025-11-02"
author: "Jessica Xia"
date: "2025-10-24"
output:
  html_document: default
  pdf_document:
    latex_engine: xelatex
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data Processing

```{r reset env}
rm(list = ls())
source("00_requirements.R")
```

First we will load the data:

```{r datasets}
#events in US from 1980 - 2024
events_data <- read.csv("data/events-US-1980-2024-Q4.csv")
events_data <- na.omit(events_data)
events_data <- events_data[-c(1, 2), ]

#insurance data
overall_insurance <- read.csv("data/all_insurance_costs.csv")
overall_insurance <- na.omit(overall_insurance)

#property insurance cost data
prop_insurance <- read.csv("data/property_insurance_costs.csv")
prop_insurance <- na.omit(prop_insurance)
```

Next we will fix the data types of our data

```{r}
colnames(events_data) <- c("Name_Date", #chr
                           "Disaster_Type", #chr
                           "Begin_Date", #num -> chr (date format)
                           "End_Date", #num -> chr (date format)
                          "CPI_Adjusted_Cost_Millions", #num
                          "Unadjusted_Cost_Millions", #num
                          "Deaths") #int

events_data$CPI_Adjusted_Cost_Millions <- as.numeric(events_data$CPI_Adjusted_Cost_Millions)
events_data$Unadjusted_Cost_Millions <- as.numeric(events_data$Unadjusted_Cost_Millions)
events_data$Deaths <- as.integer(events_data$Deaths)


events_data$Begin_Date <- as.Date(as.character(events_data$Begin_Date),
                          format = "%Y%m%d")

events_data$End_Date <- as.Date(as.character(events_data$End_Date), 
                        format = "%Y%m%d")

events_data$CPI_Adjusted_Cost_Millions <- as.numeric(events_data$CPI_Adjusted_Cost_Millions)

events_data$Unadjusted_Cost_Millions <- as.numeric(events_data$Unadjusted_Cost_Millions)

events_data$Deaths <- as.numeric(events_data$Deaths)

head(events_data)
dim(events_data)
```

```{r}
colnames(overall_insurance) <- c("PPI_Series_ID", 
                              "Year", 
                              "Month_Code", 
                              "Time_Period", 
                              "All_Insurance_Index")

head(overall_insurance)

```

```{r}
colnames(prop_insurance) <- c("Series_ID", 
                  "Year", 
                  "Month_Code", 
                  "Time_Period", 
                  "Property_Insurance_Index")

head(prop_insurance)
```

```{r savedata}
save.image("12_cleanedData.RData")
```
