==> 21_DataProcessing.Rmd <==
---
title: "RMD_21_DataProcessing"
output: html_document
date: "2025-11-04"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data Processing

```{r reset env}
rm(list = ls())
source("00_requirements.R")
```

```{r}
load("12_cleanedData.RData")
head(events_data)
head(overall_insurance)
head(prop_insurance)
```

Our overall insurance data does not contain the months of March 2009 to May 2009, so we will remove them from the property insurance data so they can match 1 to 1

```{r matchprop}
prop_insurance <- prop_insurance[-c(1:3),]
```

We can now merge the 2 datasets, since they match 1 to 1 in terms of date. We can also get rid of useless columns

```{r datamerge}

all_insurance <- merge(prop_insurance, overall_insurance)
head(all_insurance)
#we can now get rid of all and prop insurance
rm(overall_insurance)
rm(prop_insurance)
```

```{r columnremoval}

all_insurance <- all_insurance[ -c(3,4,6)]
head(all_insurance)


```

## Finding the cost of each month

```{r}
head(events_data)
```

All the events have a begin date and an end date. For simplicity, we will be assuming the cost of one of the disasters is split evenly over the months in which it occured. To do so, since we don't care about the specific days, just the months and years, first we will floor all the dates to make them easier to work with

```{r}

events_data$Begin_Date <- floor_date(events_data$Begin_Date, "month")
events_data$End_Date <- floor_date(events_data$End_Date, "month")
head(events_data)
```

Next, we will get the number of months between the begin and end dates. Add one (to account for the starting month), and that will be the duration of our event (in months). We can then find the cost per month

(I got the at period code off stackexchange)

```{r}
events_data$Duration_Interval <- interval(events_data$Begin_Date, events_data$End_Date)

events_data$Duration_Months <- as.period(events_data$Duration_Interval
) %/% months(1) + 1
#events_data$Duration_Months

events_data$Adjusted_CPM_Millions <- events_data$CPI_Adjusted_Cost_Millions / events_data$Duration_Months

events_data$Unadjusted_CPM_Millions <- events_data$Unadjusted_Cost_Millions / events_data$Duration_Months

head(events_data)
```

We can use this data to view the relative cost of disasters - perhaps the longer lasting ones are less destructive per month or equally as so. We will do so in the preliminary visualization. However, now we have to get the cost of each disaster into the insurance data. First, we will remove the M from the month code so we have a year and month, then we can turn that into a date

```{r}
all_insurance$Month <- substr(all_insurance$Month_Code, 2, 3)
all_insurance$Calc_Date <- as.Date(
  paste(all_insurance$Year, all_insurance$Month, "1", sep = "-")
)
head(all_insurance)

```

Now, we can compare the calculated date to the interval in the events_data in order to see if that month incurred the cost of the event.

```{r}
all_insurance$Disaster_Cost <- 0

A <- all_insurance[,c(6:7)] #get date
B <- events_data[,c(3,4,10)] #get date interval and disaster cost
#idk if this is neccessary but its the example on the rdocumentation website so ill go with what i know works https://www.rdocumentation.org/packages/data.table/versions/1.17.8/topics/foverlaps 
A$start <- A$Calc_Date
A$end <- A$Calc_Date
B$start <- B$Begin_Date
B$end <- B$End_Date
B <- data.table(B)
A <- data.table(A)
setkey(B, start, end)
overlaps <- foverlaps(A, B, type = "any", which = TRUE)
overlaps <- data.frame(overlaps)
for (i in 1:nrow(overlaps)) {
  monthindex <- overlaps[i,1]
  costindex <- overlaps[i,2]
  all_insurance[monthindex,]$Disaster_Cost <- all_insurance[monthindex,]$Disaster_Cost + B[costindex,]$Adjusted_CPM_Millions
}
rm(A, B, i, costindex, monthindex)
head(all_insurance)
#i testesd this with an inefficeint brute force alg and it matches up, although leaves N/A where the cost is 0, thats fine
```

Now we can clean up (we're keeping overlaps in case its useful later)

```{r}
#we can use the month date instead with month(date) and year(date) 
all_insurance$Year <- NULL
all_insurance$Month <- NULL
all_insurance$Month_Code <- NULL
all_insurance <- all_insurance[, c("Calc_Date", "Disaster_Cost", "Property_Insurance_Index", "All_Insurance_Index")]
colnames(all_insurance)[1] <- "Insurance_Month"

#if the sum was 0 the cost is n/a, turn that back into 0
#taken from tidyverse replace na reference
all_insurance$Disaster_Cost <- replace_na(all_insurance$Disaster_Cost, 0)

head(all_insurance)

events_data$Duration_Interval <- NULL
colnames(events_data)[3] <- "Begin_Month"
colnames(events_data)[4] <- "End_Month"
head(events_data)
```

```{r savedata}
save.image("22_processedData.RData")

```
